Refactoring some code xdptdr
J'ai fait plein de trucs c'est trop b1
